+++ kurtosis-module/impl/participant_network/cl/lodestar/lodestar_launcher.go
@@ -53,6 +53,7 @@ var lodestarLogLevels = map[module_io.ParticipantLogLevel]string{
 	module_io.ParticipantLogLevel_Warn:  "warn",
 	module_io.ParticipantLogLevel_Info:  "info",
 	module_io.ParticipantLogLevel_Debug: "debug",
+	module_io.ParticipantLogLevel_Trace: "silly",
 }
 
 type LodestarClientLauncher struct {