+++ kurtosis-module/impl/participant_network/cl/nimbus/nimbus_launcher.go
@@ -63,6 +63,7 @@ var nimbusLogLevels = map[module_io.ParticipantLogLevel]string{
 	module_io.ParticipantLogLevel_Warn:  "WARN",
 	module_io.ParticipantLogLevel_Info:  "INFO",
 	module_io.ParticipantLogLevel_Debug: "DEBUG",
+	module_io.ParticipantLogLevel_Trace: "TRACE",
 }
 
 type NimbusLauncher struct {