+++ kurtosis-module/impl/participant_network/cl/prysm/prysm_launcher.go
@@ -74,6 +74,7 @@ var prysmLogLevels = map[module_io.ParticipantLogLevel]string{
 	module_io.ParticipantLogLevel_Warn:  "warn",
 	module_io.ParticipantLogLevel_Info:  "info",
 	module_io.ParticipantLogLevel_Debug: "debug",
+	module_io.ParticipantLogLevel_Trace: "trace",
 }
 
 type PrysmCLClientLauncher struct {